+++ lib/Proxy/YourlsProxy.php
@@ -65,8 +65,8 @@ class YourlsProxy extends AbstractProxy
      */
     protected function _extractShortUrl(array $data): ?string
     {
-        if (($data['statusCode'] ?? 0) === 200) {
-            return $data['shorturl'] ?? 0;
+        if ((int) ($data['statusCode'] ?? 0) === 200) {
+            return $data['shorturl'] ?? null;
         }
         return null;
     }